moved twitch data source classes to each files
