+++ lib/src/main/java/com/freshdigitable/yttt/TwitchOauthViewModel.kt
@@ -6,7 +6,7 @@ import androidx.lifecycle.liveData
 import androidx.lifecycle.viewModelScope
 import com.freshdigitable.yttt.data.AccountRepository
 import com.freshdigitable.yttt.data.TwitchLiveRepository
-import com.freshdigitable.yttt.data.TwitchOauthToken
+import com.freshdigitable.yttt.data.model.TwitchOauthToken
 import com.freshdigitable.yttt.data.model.TwitchUser
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.SharingStarted
