+++ lib/src/main/java/com/freshdigitable/yttt/compose/TwitchOauthScreen.kt
@@ -17,7 +17,7 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import com.bumptech.glide.integration.compose.ExperimentalGlideComposeApi
 import com.freshdigitable.yttt.TwitchOauthViewModel
 import com.freshdigitable.yttt.compose.preview.LightModePreview
-import com.freshdigitable.yttt.data.TwitchOauthToken
+import com.freshdigitable.yttt.data.model.TwitchOauthToken
 import com.freshdigitable.yttt.data.model.TwitchUser
 import kotlinx.coroutines.launch
