+++ repository-twitch/src/main/java/com/freshdigitable/yttt/data/model/TwitchOauthToken.kt
@@ -1,0 +1,10 @@
+package com.freshdigitable.yttt.data.model
+
+data class TwitchOauthToken(
+    val accessToken: String,
+    val scope: String,
+    val state: String,
+    val tokenType: String,
+) {
+    companion object
+}
