+++ repository-twitch/src/main/java/com/freshdigitable/yttt/data/source/remote/TwitchOauthService.kt
@@ -1,0 +1,19 @@
+package com.freshdigitable.yttt.data.source.remote
+
+import okhttp3.ResponseBody
+import retrofit2.Call
+import retrofit2.http.GET
+import retrofit2.http.Query
+import java.util.UUID
+
+internal interface TwitchOauthService {
+    /// https://dev.twitch.tv/docs/authentication/getting-tokens-oauth/
+    @GET("oauth2/authorize?response_type=token")
+    fun authorizeImplicitly(
+        @Query("client_id") clientId: String,
+        @Query("force_verify") forceVerify: Boolean? = null,
+        @Query("redirect_uri") redirectUri: String,
+        @Query("scope") scope: String,
+        @Query("state") state: String = UUID.randomUUID().toString(),
+    ): Call<ResponseBody>
+}
